bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
# pydantic==2.7.0
python-dotenv==1.0.0
redis==5.0.1
google-cloud-firestore==2.13.1
